### 2019-09-19 #1750 @danielskatz
but just if the paper is a different .md - a .rmd will fail...
