### 2019-09-19 e221960
renamed "tag" to "tags"
in case it helps incrementally with the whedon error message
