### 2019-09-19 #1750 @lskatz
I fixed it by making `tag`, `author`, and `affiliation` plural in the yaml.
