### 2019-09-20 4325c73
updated CPAN installation to the more basic `-MCPAN -e shell`
