### 2019-09-23 #1750 @csoneson
@retimme, @ondovb, @aphillippy, @rvosa, @tseemann, @iimog - would one of you be interested in reviewing this submission for JOSS?
