### 2019-09-25 #1762 @whedon
[REVIEW]: Mashtree: a rapid comparison of whole genome sequence files
**Submitting author:** @lskatz (<a href="http://orcid.org/0000-0002-2533-9161">Lee Katz</a>)
**Repository:** <a href="https://github.com/lskatz/mashtree" target ="_blank">https://github.com/lskatz/mashtree</a>
**Version:** v1.0
