### 2019-09-25 #1762 @csoneson
@lskatz - please have a look at the checklists above and make sure that your submission has all required components. Otherwise, please add them already now to simplify the review.

A couple of comments about your paper that would be good if you could address:
- it's a bit on the long side for JOSS standards (papers are typically between 250 and 1,000 words).
- your figures should have informative legends (currently they're only "Figure 1" etc.)
