### 2019-09-25 #1762 @lskatz
Thank you @luizirber and @peterwc for reviewing!

Thank you @csoneson!  I edited the figure legends so that they have titles.  I also took the text that I thought was in the figure legend and placed them into the actual legend.  I will ask the bot to regenerate the pdf to make sure the formatting works.  I can also remove some text from the introduction in a future edit.
