### 2019-09-25 75ebda9
figure legends alteration
