+++ paper/paper.md
@@ -1,6 +1,6 @@
 ---
 title: 'Mashtree: a rapid comparison of whole genome sequence files'
-authors:
+author:
 - affiliation: "1, 2"
   name: Lee S. Katz
   orcid: 0000-0002-2533-9161
@@ -29,13 +29,13 @@ output:
   pdf_document: default
   word_document: default
 bibliography: paper.bib
-tags:
+tag:
 - dendrogram
 - mash
 - sketch
 - tree
 - rapid
-affiliations:
+affiliation:
 - index: 1
   name: Enteric Diseases Laboratory Branch, Centers for Disease Control and Prevention,
     Atlanta, GA, USA
@@ -109,57 +109,35 @@ The Mashtree package is programmed in Perl, and is available in the CPAN reposit
 
 # Figures
 
-![Figure 1](Mashtree_workflow.png)  
-
-## Figure 1.  The Mashtree workflow
-
-### Step 1
-
+![Figure 1. The Mashtree workflow.  
+Step 1)
 Sketch genomes with Mash. In this schematic, there is a green circle representing each genome in the analysis.  Filled-in brown circles indicate the presence of a kmer.  Missing circles represent true absence.  After hashing with a sketch size of six (after the arrow), some kmers are not represented in the Mash sketch either because they are not present in the original genome or because only a finite number of kmers are sketched (e.g., six in this example).  Henceforth, truly missing hashes or hashes not included in the Mash sketch are represented by empty circles.  
-
-### Step 2
-
+Step 2)
 Calculate distances with `Mash dist`.  Distances in the figure are represented by Jaccard distances, which are calculated as the intersection divided by the union.  In this example, the genomes are separated by Jaccard distances of 5/9, 4/9, and 3/9.  These jaccard distances are internally transformed into Mash distances [@Ondov:2016].  
+Step 3)
+Create dendrogram with Quicktree using the Mash distance matrix.  
+](Mashtree_workflow.png)  
 
-### Step 3.
 
-Create dendrogram with Quicktree using the Mash distance matrix.
-
-![Figure 2](Bootstrap_workflow.png)  
-
-## Figure 2 The Mashtree bootstrap workflow
-
-### Step 1
-
-Generate a tree with the normal workflow [Figure 1](Figure 1). This is the main tree.
-
-### Step 2
-
+![Figure 2. The Mashtree bootstrap workflow.  
+Step 1)
+Generate a tree with the normal workflow [Figure 1](Figure 1). This is the main tree.  
+Step 2)
 Run the normal workflow once per replicate but with a different random seed. In this example, the top right replicate differs from the main tree.  All ten of these trees are the bootstrap tree replicates.  
+Step 3)
+For each parent node in the main tree, quantify how many bootstrap tree replicates have the same node with the same children. Record that percentage next to each parent node. This percentage quantifies how confident the Mashtree cluster is, controlling for the random seed in the Mash program.  
+](Bootstrap_workflow.png)  
 
-### Step 3
-
-For each parent node in the main tree, quantify how many bootstrap tree replicates have the same node with the same children. Record that percentage next to each parent node. This percentage quantifies how confident the Mashtree cluster is, controlling for the random seed in the Mash program.
-
-![Figure 3](Jackknife_workflow.png)  
-
-## Figure 3.  The Mashtree jackknife workflow
-
-### Step 1 
-
-Generate a tree with the normal workflow [Figure 1](Figure 1). This is the main tree.
-
-### Step 2
-
+![Figure 3.  The Mashtree jackknife workflow.  
+Step 1)
+Generate a tree with the normal workflow [Figure 1](Figure 1). This is the main tree.  
+Step 2)
[…]
