### 2019-09-25 bed2778
added back in tags, authors, affiliations tags
