### 2019-09-25 6227991
figure legends on one line each
