### 2019-09-25 c744861
removed link to figure 1 from other figures, in case that is the error.
