### 2019-09-26 c359f1f
removed innacurate statement that mashtree_bootstrap only works on compressed fastq files
