+++ bin/mashtree_bootstrap.pl
@@ -235,8 +235,6 @@ sub usage{
   --outmatrix          ''   Output file for distance matrix
   --reps               0    How many bootstrap repetitions to run;
                             If zero, no bootstrapping.
-                            Bootstrapping will only work on compressed fastq
-                            files.
   --numcpus            1    This will be passed to mashtree and will
                             be used to multithread reps.
   
