### 2019-09-26 2e66360
removed innacurate statement that mashtree_bootstrap only works on compressed fastq files
