### 2019-10-16 #1762 @peterwc
Working on it this week. I apologize for the delay. Expecting a Friday
completion.

Thanks,
Peter

On Thu, Oct 10, 2019 at 10:58 AM Charlotte Soneson <notifications@github.com>
wrote:

> Hi @luizirber <https://github.com/luizirber>, @peterwc
> <https://github.com/peterwc> - just checking in on how the reviews are
> progressing. Do you think you can have comments for @lskatz
> <https://github.com/lskatz> within the next week or so? Don't hesitate to
> let me know if you have questions. Thanks!
>
