### 2019-10-16 #1762 @luizirber
Overall this is all very good, thanks @lskatz! Here's some review points:

## Functionality

### Installation

The install instructions and requirements could have a bit more information, I
had to dig a bit to find proper Perl package names (for use with `cpanm`) and
links to `quicktree` and `mash`.

Also adding that mashtree is available on bioconda!

This is what I ended up doing to install using an Ubuntu docker image and the
"Installation from Git" section:

```
git clone https://github.com/lskatz/mashtree.git
cd mashtree
docker run -it -v $PWD:/mashtree ubuntu:latest /bin/bash
```

Inside docker (travis config was more useful here):
https://github.com/lskatz/mashtree/blob/e9697e7496b5c6a85ae69832189ea8350cacfdfb/.travis.yml

```
apt update
apt install -y build-essential cpanminus libexpat1-dev wget sqlite3 libsqlite3-dev

cd /mashtree

cpanm -n BioPerl Bio::Sketch::Mash DBD::SQLite DBI Graph::Dijkstra

wget https://github.com/marbl/Mash/releases/download/v2.2/mash-Linux64-v2.2.tar
tar xf mash-Linux64-v2.2.tar
mv mash-Linux64-v2.2/mash /usr/bin/mash

wget https://github.com/khowe/quicktree/archive/v2.5.zip
tar xf v2.5.tar.gz
cd quicktree-2.5
make
mv quicktree /usr/bin
cd ..

perl Makefile.PL
make test
```

## Documentation

### Installation instructions

(see above)

### Functionality documentation

This is missing for using it as a library in Perl, but since the primary form of
interaction is through the CLI I think it is OK.

### Community guidelines

I didn't see it in the README, but you can also create a CONTRIBUTING file
(GitHub recognizes it and render the content when a new issue or PR is opened):
https://github.blog/2012-09-17-contributing-guidelines/

## Software paper

### References

Most of them are missing DOIs, can you please add them?
