### 2019-10-17 d63a82b
INSTALLATION and CONTRIBUTING guidelines
