### 2019-10-17 d019249
changed some italics to bold
