### 2019-10-17 #1762 @luizirber
@lskatz put the DOIs in the paper.bib file in your repo, when the paper is regenerated by whedon they will be picked up.
