+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={16185},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nmicrobiol.2016.185}
 }
 
 @article{Souvorov:2018,
@@ -17,7 +18,8 @@
   number={1},
   pages={153},
   year={2018},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.5281/zenodo.1407162}
 }
 
 @article{Ondov:2016,
@@ -28,7 +30,8 @@
   number={1},
   pages={132},
   year={2016},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s13059-016-0997-x}
 }
 
 @article{Saitou:1987,
@@ -38,7 +41,8 @@
   volume={4},
   number={4},
   pages={406--425},
-  year={1987}
+  year={1987},
+  doi={10.1093/oxfordjournals.molbev.a040454}
 }
 
 @article{Howe:2002,
@@ -49,7 +53,8 @@
   number={11},
   pages={1546--1547},
   year={2002},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/18.11.1546}
 }
 
 @article{Harris:2018,
@@ -58,7 +63,8 @@
   journal={BioRxiv},
   pages={453142},
   year={2018},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/453142}
 }
 
 @article{Gardner:2015,
@@ -69,7 +75,8 @@
   number={17},
   pages={2877--2878},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btv271}
 }
 
 @article{Timme:2017,
@@ -79,18 +86,8 @@
   volume={5},
   pages={e3893},
   year={2017},
-  publisher={PeerJ Inc.}
-}
-
-@article{Oakeson:2017,
-  title={Bioinformatic analyses of whole-genome sequence data in a public health laboratory},
-  author={Oakeson, Kelly F and Wagner, Jennifer Marie and Mendenhall, Michelle and Rohrwasser, Andreas and Atkinson-Dunn, Robyn},
-  journal={Emerging infectious diseases},
-  volume={23},
-  number={9},
-  pages={1441},
-  year={2017},
-  publisher={Centers for Disease Control and Prevention}
+  publisher={PeerJ Inc.},
+  doi={10.7717/peerj.3893}
 }
 
 @article {Baker:2019,
@@ -167,7 +164,8 @@
   number={7},
   pages={441--450},
   year={2019},
-  publisher={Mary Ann Liebert, Inc., publishers 140 Huguenot Street, 3rd Floor New~.}
+  publisher={Mary Ann Liebert, Inc., publishers 140 Huguenot Street, 3rd Floor New~.},
+  doi={10.1089/fpd.2019.2662}
 }
 
 @article{Alikhan:2018,
@@ -178,7 +176,8 @@
   number={4},
   pages={e1007261},
   year={2018},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pgen.1007261}
 }
 
 @article{Maiden:2013,
@@ -189,5 +188,6 @@
   number={10},
   pages={728},
   year={2013},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nrmicro3093}
 }
