### 2019-10-18 #1762 @lskatz
Thank you for your comments @luizirber !  I did the following:

1.  Added a committed installation document at https://github.com/lskatz/mashtree/blob/master/docs/INSTALL.md
  * This document is linked from the README, https://github.com/lskatz/mashtree/blob/master/README.md#installation
  * All instructions you suggested have been incorporated.  They were very helpful, thank you!  However, I switched some details so that it could be installed outside of Docker.  Docker at my institution and at many other places is a security risk and so I feel like installation in a home directory is more universal.
2. Functionality documentation:  yes, I believe that since it is offered as a package in Perl that the most appropriate mode of distribution was in the native package manager, CPAN.  It is not meant strictly as a library but as a CLI executable.  So I think your train of thought is correct.
  * I am very flattered that Mashtree has been incorporated in package managers, and so I made some other links here with a disclaimer that I did not create them: https://github.com/lskatz/mashtree/blob/master/docs/INSTALL.md#other-sources-for-mashtree
3. Thank you for the community guidelines suggestion.  I added guidelines per your suggestion.
  * Linked from here: https://github.com/lskatz/mashtree/blob/master/README.md#further-documentation
  * CONTRIBUTING.md here: https://github.com/lskatz/mashtree/blob/master/CONTRIBUTING.md
4.  I added DOIs and will double check whether @whedon generates the PDF without error momentarily.
