### 2019-10-20 #1762 @peterwc
@csoneson I will continue with my review but I realized that I may be in a gray area for this review. I apologize for not bringing it up sooner. I am a bioinformatics researcher at the CDC, but I work in a different division than @lskatz. I have not been on any papers with him, but I do know him personally. After reading through the ethics guidelines I decided it would be best to ask for waiver before submitting my review. I am impartial and will provide a complete and thorough review if I am granted a waiver.
