### 2019-10-21 #1762 @csoneson
@peterwc - thanks a lot for your disclosure. After consideration, and in light of your personal connection with @lskatz, I would actually prefer to try to recruit a different reviewer for this submission; but I would like to thank you for engaging with JOSS and I hope we can get back to you for another review in the future!
