### 2019-10-21 #1762 @csoneson
@ondovb - would you be interested in reviewing this submission for the Journal of Open Source Software?

"Mashtree: a rapid comparison of whole genome sequence files"
GitHub repository: https://github.com/lskatz/mashtree
