### 2019-10-23 #1762 @csoneson
@standage, @greatfireball - that's great, thank you for agreeing to review, it would be great to get your input! I'll add you both as reviewers here, and also add a checklist for each of you in the first post above.
