### 2019-10-23 #1762 @luizirber
>     * All instructions you suggested have been incorporated.  They were very helpful, thank you!  However, I switched some details so that it could be installed outside of Docker.  Docker at my institution and at many other places is a security risk and so I feel like installation in a home directory is more universal.

Sounds great! I did in docker to avoid having to install a lot of extra packages in my system, but doing a home directory installation is the best path (since many people won't have admin rights to their system).

>     1. Thank you for the community guidelines suggestion.  I added guidelines per your suggestion.
>     1. I added DOIs and will double check whether @whedon generates the PDF without error momentarily.

This all looks good. Thanks @lskatz!

@csoneson, checklist complete for me.
