### 2019-10-29 49131c2
INSTALL.md Quicktree unzipping cmd error  #46
Fixed wget/tar inconsistency
