### 2019-10-31 #1762 @standage
```
---
title: 'Review of Mashtree: a rapid comparison of whole genome sequence files'
author: Daniel Standage
date: 2019-10-31
---
```

The authors describe Mashtree, a utility for constructing trees from sequence reads or assemblies using Min-hash sketch distances. Mashtree leverages Mash to compute distances rapidly and efficiently, and QuickTree to build a neighbor-joining tree from the resulting distance matrix. Mashtree supports two modes of operation: a "fast" mode based on a fixed minimum *k*-mer abundance filter, and an "accurate" mode where the filter is inferred empirically from the *k*-mer abundance distribution. Mashtree also implements two resampling methods for estimating the confidence of the trees it constructs.

Mashtree provides several features bioinformatics and genomics practitioners will find convenient, including the following.

- Seamless support for both compressed and uncompressed sequence data.
- Support for pre-computed Min-hash sketches, and for saving sketches computed by Mashtree for subsequent use.
- Support for multi-threaded execution to take advantage of available system computing capacity.

I have added several comments below for the authors' consideration, but in any case I endorse this submission for publication in JOSS.

## Installation

**TL;DR** Conda and from-source installation modes seem to work well. 👍

Mashtree's documentation describes multiple modes of installation/execution, including Bioconda, Docker, CPAN, and from the source code distribution. Installing from Bioconda on my OS X system was painless, although I was skeptical whether this reflected the latest version. Accordingly, I also installed from source (latest `master` on Githuub) on a fresh Ubuntu VM. This process was also fairly straightforward, although I do have a couple comments.

- The [INSTALL.md](https://github.com/lskatz/mashtree/blob/49131c2db7dc3433c441bf910864ab0a26836502/docs/INSTALL.md) document suggests `export PERL5LIB=$PERL5LIB:$HOME/lib/perl5` for configuring a Perl environment in your home directory. I'm not sure whether I missed a step, but initially Mashtree couldn't find the requisite libraries I had just installed. After tinkering for a bit, I was able to finally able to get it to work by adding `$HOME/perl5/lib/perl5` to the `PERL5LIB` variable. Although not necessary for this review, I suggest that the authors try the installation procedures verbatim on a fresh Linux VM and update the documentation if/as needed.
- Tools like virtualenv and (more recently) Conda have spoiled me when it comes to configuring isolated environments for testing and developing software. Installing the latest version of Mashtree from Bioconda will suffice in most circumstances. But if I wanted to test a version of Mashtree not yet available from Bioconda, or set up a development environment where I could tinker with Mashtree, I'm not entirely sure how to do so in a way that doesn't scatter library detritus all over my filesystem. I'm guessing that setting up a dedicated `perl5/` instance somewhere (and setting `PERL5LIB`  appropriately) would work, although I'm not 100% certain I could go back later to track down and remove all library installation artifacts. I also suspect that installing Mashtree in a dedicated Conda environment (to handle all prerequisites) and then running Mashtree directly from a git clone of the latest source would also work. **<u>NOTE</u>: This isn't a critique of Mashtree, nor even a firm suggestion to the authors necessarily.** I just wanted to point out a potential use case for their consideration.

## Best Practices

**TL;DR** CI GOOD 👍, MOAR PR!

The Mashtree repository includes an automated test suite, which is executed frequently in a Contintuous Integration (CI) environment. Potential users of Mashtree can be confident that bugs will be identified early and often, and hopefully most will be addressed before stable releases are published.

Pull requests and CI have been used to review 3rd-party (and possibly internal/collaborator) contributions. I'd suggest that the corresponding author consider using pull requests more routinely—i.e., for all non-trivial updates to the repository. Even for projects where I am the sole active contributor, I have found value in providing a summary and description of each update, linking to relevant issue threads as necessary, and waiting to merge the pull request until the CI build passes.
