### 2019-10-31 #1762 @standage
@csoneson, this completes my review of Mashtree.

@lskatz, if you have any comments or questions, feel free to keep the conversation going here or on the Mashtree issue tracker—just @ me! 😎
