### 2019-10-31 #1762 @csoneson
> @csoneson, this completes my review of Mashtree.

@standage - thanks for this!
