### 2019-11-01 #1762 @lskatz
Thank you @standage for that wonderful review!  To address one concern, I think that the issue might have been that I did not have a complete `cpanm` command, when I should have had `-l ~` in there.  So now one of the lines in INSTALL.md now reads as

    cpanm -l ~ --notest BioPerl Bio::Sketch::Mash DBD::SQLite DBI Graph::Dijkstra

Fixed in 71f7470
