+++ README.md
@@ -101,9 +101,13 @@ For more information on contributions, please see [CONTRIBUTING.md](CONTRIBUTING
 
 ## Citation
 
-The paper is in preparation but for now, this is a valid citation:
+### In review, JOSS
 
-Katz, L. S., Griswold, T., & Carleton, H. A. (2017, October 8-11). [_Generating WGS Trees with Mashtree_](misc/mashtree%20ASM%20NGS.pptx). Poster presented at the American Society for Microbiology Conference on Rapid Applied Microbial Next-Generation Sequencing and Bioinformatic Pipelines, Washington, DC. 
+Katz, L. S., Griswold, T., Morrison, S., Caravas, J., Zhang, S., den Bakker, H.C., Deng, X., and Carleton, H. A., "Mashtree: a rapid comparison of whole genome sequence files."  Journal of Open Source Software, in review.  https://github.com/openjournals/joss-reviews/issues/1762#issuecomment-543736907
+
+### Poster
+
+Katz, L. S., Griswold, T., & Carleton, H. A. (2017, October 8-11). [_Generating WGS Trees with Mashtree_](misc/mashtree%20ASM%20NGS.pptx). Poster presented at the American Society for Microbiology Conference on Rapid Applied Microbial Next-Generation Sequencing and Bioinformatic Pipelines, Washington, DC. Poster number 27.
 
 ## GitHub stickers for Mashtree
 
