### 2019-11-27 #1762 @csoneson
@lskatz - since we have two thorough finalized reviews I'm going to move on with the processing of this paper for acceptance. As discussed previously, it is a bit long for JOSS (currently ~2,000 words, while the typical JOSS paper should be between 250 and 1,000 words). Could you try to make it a bit more concise? Please see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain for more details.
