### 2019-11-28 #1762 @lskatz
Thank you @csoneson! I will gladly shorten the paper! Can I aim for 1500 words? And separately, would it be okay if I made some sort of supplemental if I need to keep some text?
