### 2019-11-28 #1762 @csoneson
Rather than focusing on the exact number of words to target, I'd suggest to:
* summarize the existing approaches a bit more concisely
* put a short, high-level statement (1-2 sentences perhaps) of what Mashtree does/the purpose of the software earlier in the Summary. Right now, one has to read more than one page before it is mentioned.
* focus on the research purpose and functionality of Mashtree. Details on the implementation and different options for running the software are more suitable as part of the documentation, in the software repository.
