### 2019-12-02 #1762 @lskatz
Thank you!  One more question:  How do I avoid the syntax happening right now in the figure legends?  For example:

> **Figure 1:** Figure 1.
