### 2019-12-02 #1762 @csoneson
I think you can remove the `Figure 1. ` part from your figure legend in the .md file.
