### 2019-12-02 b8c1e75
removed duplicate Figure X syntax from figure legends
