### 2019-12-02 0115713
reduced some introduction
