### 2019-12-02 0bc9a83
tried to fix Figure 1 callout
