### 2019-12-02 f30023f
greatly reduced the intro
