+++ paper/paper.bib
@@ -191,3 +191,14 @@
   publisher={Nature Publishing Group},
   doi={10.1038/nrmicro3093}
 }
+
+@article{bloom:1970,
+  title={Space/time trade-offs in hash coding with allowable errors},
+  author={Bloom, Burton H},
+  journal={Communications of the ACM},
+  volume={13},
+  number={7},
+  pages={422--426},
+  year={1970},
+  publisher={ACM}
+}
