### 2019-12-02 c3f8fff
documentation on the algorithms
