+++ docs/ALGORITHM.md
@@ -1,0 +1,29 @@
+# Algorithms
+
+## Fast or accurate mode
+
+While Mashtree has been developed for speed, some might want higher accuracy.
+In response, we have implemented an optional minimum abundance filter for raw read input files.
+This filter creates a histogram of kmers using Mash and detects the valleys between peaks.
+The first (left-most) peak of a kmer histogram represents kmers that occur only once or rarely.
+These "rare" kmers usually represent sequencing errors.
+Therefore, the minimum abundance filter can optionally be used to see where the first valley is and automatically set that value for the minimum kmer coverage parameter.
+Using this filter helps avoid noise in the input data.
+Running in accurate mode is shorthand for running Mashtree with the minimum abundance filter turned on.
+Otherwise by default, a user is running in fast mode.
+
+## Confidence values
+
+Most phylogenetic trees have confidence scores associated with each hypothetical ancestor node.
+These support scores are typically generated by resampling the input data many times, analyzing each replicate in context of the original phylogeny.
+Strong  support for a node in the input data will result in consistent recovery of that node in the majority of resampled replicate trees.  Conversely, nodes which have low support in the input data will be sensitive to stochastic changes in the resampled sketches which can result in alternate topologies of involved isolates.  Therefore, the support value correlates with the robustness and reproducibility of a recovered relationship.  As the NJ tree recovered from a single tree reconstruction will always be a fully resolved binary tree, the addition of these resampled support values provides a valuable tool for assessing the confidence and reliability of a node.  This can be especially helpful when trying to assess the exact topology of clades containing  three or more very closely related genomes.
+
+Although Mashtree does not infer phylogeny, we have borrowed the ideas behind phylogenetic confidence values to yield confidence values for each parent node in the tree. There are two resampling methods implemented in Mashtree to assign support values to infernal nodes, bootstrapping and jackknifing. Initially, both methods create a tree as depicted in [Figure 1](../paper/Mashtree_workflow.png).
+
+The bootstrapping method infers confidence on the tree by creating one new tree per replication [Figure 2](../paper/Bootstrap_workflow.png). However, instead of accepting the default seed value for Mash, a new one is generated.
+Changing this seed value will affect the value of hashed kmers and how they sort, therefore providing a different sampling of kmers in the sketch.
+Due to changes in sketch sampling, distances between genomes can differ and the resulting tree in the replication can differ. Then, nodes in the original tree are assigned support values based on how frequently the same bipartition was recovered in the replicate trees.
+
+In the jackknife method, for each replication, each genome is picked one at a time as the query genome [Figure 3](../paper/Jackknife_workflow.png).  Its hashes are sampled without replacement for half of the original hashes.  Then, a Mash distance is recomputed between the query genome and all other genomes in the analysis.
+When the distances are recorded in the replicate's pairwise distance matrix, the average distance between each comparison is used of each query-reference genome pair, and vice versa. This new distance matrix is used for computing a NJ tree. Nodes in the original tree are assigned support values based on how frequently the same bipartition was recovered in replicate trees.
+
