### 2019-12-02 746f845
Seeing whether Figures are formatted right
