### 2019-12-02 9786c68
Seeing whether Figures are formatted right
