### 2019-12-02 ea5f580
Seeing whether Figures are formatted right
