### 2019-12-02 e65c928
Seeing whether Figures are formatted right
