### 2019-12-02 74f0b54
Seeing whether Figures are formatted right
