### 2019-12-02 c5c5d2e
removed a bunch of algorithm text
