+++ paper/paper.md
@@ -61,8 +61,8 @@ One major kmer strategy is to reduce each genome to split kmers. With split kmer
 When comparing two or more genomes, the variable sites are compared. Split kmers have been implemented in software packages such as KSNP and SKA [@Gardner:2015; @Harris:2018].
 
 Another major kmer strategy is to convert genomic data into manageable datasets, usually called sketches [@Ondov:2016; @Baker:2019; @Zhao:2018].
-Most notably, an algorithm called min-hash was implemented in the Mash package [@Ondov:2016].  In the min-hash algorithm, all kmers are recorded and transformed into integers using hashing and a Bloom filter [@bloom:1970].  These hashed kmers are sorted and only the first several kmers are retained.  The kmers that appear at the top of the sorted list are collectively called the sketch.  Comparing multiple sketches can be used to estimate relatedness.
-In other words, two genomes that have been sketched with the min-hash algorithm can be compared by counting how many hashed kmers they have in common. 
+Most notably, an algorithm called min-hash was implemented in the Mash package [@Ondov:2016].  In the min-hash algorithm, all kmers are recorded and transformed into integers using hashing and a Bloom filter [@bloom:1970].  These hashed kmers are sorted and only the first several kmers are retained.  The kmers that appear at the top of the sorted list are collectively called the sketch.
+Any two sketches can be compared by counting how many hashed kmers they have in common. 
 
 Because min-hash creates distances between any two genomes,
 min-hash values can be used to rapidly cluster genomes into trees using the neighbor-joining algorithm [@Saitou:1987].
@@ -70,36 +70,31 @@ We implemented this idea in software called Mashtree, which quickly and efficien
 
 # Implementation
 
-## Base workflow
+## Workflow
 
 Mashtree builds on two major algorithms that are already implemented in other software packages. The first is the min-hash algorithm, which is implemented in the software Mash [@Ondov:2016].  Mashtree uses Mash to create sketches of the genomes with the function `mash sketch`.  We elected to keep most default Mash parameters but increased the sketch size (number of hashed kmers) from 1,000 to 10,000 to increase discriminatory power.  Then, Mash is used to calculate the distances between genomes with `mash dist`.  Mashtree records these distances into a pairwise distance matrix. Next, Mashtree calls the neighbor-joining (NJ) algorithm which is implemented in the software QuickTree [@Howe:2002].  The Mash distance matrix is used with QuickTree with default options to generate a dendrogram.  The workflow is depicted in Figure 1.
 
-## Fast or accurate mode
-
-While Mashtree has been developed for speed, some might want higher accuracy.  In response, we have implemented an optional minimum abundance filter for raw read input files.  This filter creates a histogram of kmers using Mash and detects the valleys between peaks.  The first (left-most) peak of a kmer histogram represents kmers that occur only once or rarely.  These "rare" kmers usually represent sequencing errors.  Therefore, the minimum abundance filter can optionally be used to see where the first valley is and automatically set that value for the minimum kmer coverage parameter.  Using this filter helps avoid noise in the input data.  Running in accurate mode is shorthand for running Mashtree with the minimum abundance filter turned on.  Otherwise by default, a user is running in fast mode.
-
 ## Confidence values
 
-Most phylogenetic trees have confidence scores associated with each hypothetical ancestor node.
-These support scores are typically generated by resampling the input data many times, analyzing each replicate in context of the original phylogeny.
[…]
