### 2019-12-02 09c2980
fixed some typos
