### 2019-12-02 #1762 @lskatz
Thank you for all of your suggestions @csoneson !  I am generating the PDF for my coauthors to see and will let you know when the dust settles.  Do these changes sufficiently address your comments?  I cut down a lot of the introduction and the algorithm discussion, but I kept the figures.
