### 2019-12-02 #1762 @csoneson
Yes, I think this structure is good! Just a small comment: In the first paragraph, instead of "nucleotides of length k", perhaps use something like "nucleotide {stretches/sequences/...} of length k".
