### 2019-12-02 #1762 @csoneson
Also, please add the missing DOI above
