### 2019-12-03 fef87f7
added DOI for Bloom paper; minor edit on manuscript
