+++ paper/paper.bib
@@ -200,5 +200,6 @@
   number={7},
   pages={422--426},
   year={1970},
-  publisher={ACM}
+  publisher={ACM},
+  doi={10.1145/362686.362692}
 }
