### 2019-12-07 #1762 @csoneson
@lskatz - I sent a PR with a couple of fixes to the bibliography and a few typos. Please check, merge and ask whedon to generate a new pdf.
Then, please:
1) Check that version information is consistently reported
2) Make a tagged release of the software
3) Report the version here
4) Make a Zenodo archive. Check that the Zenodo archive has the same title and author list as the paper (you can also add authors' ORCIDs)
5) Report the Zenodo DOI here
