### 2019-12-08 d6e6eeb
fixed some deps; v1.0
