+++ .travis.yml
@@ -5,9 +5,11 @@ env: PERL_CPANM_OPT="--notest --force --skip-satisfied"
 apt:
   packages:
     - bioperl
+    - sqlite3
+    - libsqlite3-dev
 before_install:
-  - sudo apt-get update
-  - sudo apt-get install sqlite3 libsqlite3-dev
+  #- sudo apt-get update
+  #- sudo apt-get install sqlite3 libsqlite3-dev
   # bootstrap modules needed to manipulate dist
   #- "yes no | cpanm --force --notest Bio::Perl"
   - cpanm -v -L . DBD::SQLite
