+++ .travis.yml
@@ -7,6 +7,7 @@ apt:
     - bioperl
     - sqlite3
     - libsqlite3-dev
+    - libberkeleydb-perl
 before_install:
   #- sudo apt-get update
   #- sudo apt-get install sqlite3 libsqlite3-dev
