+++ .travis.yml
@@ -8,6 +8,7 @@ apt:
     - sqlite3
     - libsqlite3-dev
     - libberkeleydb-perl
+    - libcdb-file-perl
 before_install:
   #- sudo apt-get update
   #- sudo apt-get install sqlite3 libsqlite3-dev
