+++ .travis.yml
@@ -3,6 +3,7 @@ perl:
   - "5.24-shrplib"
 env: PERL_CPANM_OPT="--notest --force --skip-satisfied"
 apt:
+  update: true
   packages:
     - bioperl
     - sqlite3
