### 2019-12-08 df45e3a
moved bioperl installation to before_install
