+++ .travis.yml
@@ -5,14 +5,14 @@ env: PERL_CPANM_OPT="--notest --force --skip-satisfied"
 apt:
   update: true
   packages:
-    - bioperl
+    #- bioperl
     - sqlite3
     - libsqlite3-dev
     - libberkeleydb-perl
     - libcdb-file-perl
 before_install:
-  #- sudo apt-get update
-  #- sudo apt-get install sqlite3 libsqlite3-dev
+  - sudo apt-get update -y
+  - sudo apt-get install bioperl sqlite3 libsqlite3-dev
   # bootstrap modules needed to manipulate dist
   #- "yes no | cpanm --force --notest Bio::Perl"
   - cpanm -v -L . DBD::SQLite
