+++ .travis.yml
@@ -26,12 +26,13 @@ before_install:
 install:
   #- cpanm --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
   - export PATH=$PATH:$HOME/bin
-  - cpanm --installdeps --notest .
+  - export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
 before_script:
   - ls -F # if this test fails, just be sure what the directory is
   - PATH=$PATH:./mash-Linux64-v2.2
   - export PATH=$PATH:./quicktree
 script:
+  - cpanm --installdeps . --notest .
   - perl Makefile.PL 
   - perl bin/mashtree -h 2>&1 | grep .
   - perl bin/mashtree --help 2>&1 | grep .
