+++ .travis.yml
@@ -24,7 +24,6 @@ before_install:
   - git clone https://github.com/khowe/quicktree.git
   - make -C quicktree
 install:
-  #- cpanm --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
   - export PATH=$PATH:$HOME/bin
   - export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
 before_script:
@@ -32,6 +31,7 @@ before_script:
   - PATH=$PATH:./mash-Linux64-v2.2
   - export PATH=$PATH:./quicktree
 script:
+  - cpanm --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
   - cpanm --installdeps . --notest .
   - perl Makefile.PL 
   - perl bin/mashtree -h 2>&1 | grep .
