+++ Makefile.PL
@@ -49,9 +49,9 @@ WriteMakefile1(
     'File::Which'    => 0,
     'DBI'            => 0,
     'DBD::SQLite'    => 0,
-    'Bio::Matrix::IO'=> 0,
-    'Bio::Tree::DistanceFactory'=> 0,
-    'Bio::Tree::Statistics'=> 0,
+    'Bio::Matrix::IO'=> '==1.7.3',
+    'Bio::Tree::Statistics'=> '==1.7.3',
+    'Bio::Tree::DistanceFactory'=> '==1.7.3',
     'Bio::Sketch::Mash'    => 0, # ensures that Mash gets loaded
     #'Bio::Kmer'      => 0.19,
     #'Graph::Dijkstra'=> 0,
