+++ .travis.yml
@@ -31,10 +31,10 @@ before_script:
   - PATH=$PATH:./mash-Linux64-v2.2
   - export PATH=$PATH:./quicktree
 script:
-  - cpanm -l . --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
-  - find . -type f -name IO.pm || true
-  - find . -type f -name DistanceFactory.pm || true
-  - cpanm --installdeps . --notest .
+  #- cpanm -l . --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
+  #- find . -type f -name IO.pm || true
+  #- find . -type f -name DistanceFactory.pm || true
+  - cpanm --installdeps --notest .
   - perl Makefile.PL 
   - perl bin/mashtree -h 2>&1 | grep .
   - perl bin/mashtree --help 2>&1 | grep .
