+++ .travis.yml
@@ -11,8 +11,9 @@ apt:
     - libberkeleydb-perl
     - libcdb-file-perl
 before_install:
-  - sudo apt-get update -y
-  - sudo apt-get install bioperl sqlite3 libsqlite3-dev
+  - cpanm --local-lib=~/perl5 local::lib && eval $(perl -I ~/perl5/lib/perl5/ -Mlocal::lib)
+  #- sudo apt-get update -y
+  #- sudo apt-get install bioperl sqlite3 libsqlite3-dev
   # bootstrap modules needed to manipulate dist
   #- "yes no | cpanm --force --notest Bio::Perl"
   - cpanm -v -L . DBD::SQLite
