### 2019-12-09 edb7dc3
trying to install bioperl via specific travis command
