### 2019-12-09 e564d53
trying to install bioperl via specific travis command
