### 2019-12-09 29a17fb
trying to install bioperl via specific travis command
