+++ .travis.yml
@@ -6,6 +6,7 @@ apt:
   update: true
   packages:
     #- bioperl
+    - libgd-dev
     - sqlite3
     - libsqlite3-dev
     - libberkeleydb-perl
@@ -37,6 +38,7 @@ script:
   #- cpanm -l . --notest --force -v Bio::Tree::Statistics Bio::Matrix::IO Bio::Tree::DistanceFactory
   #- find . -type f -name IO.pm || true
   #- find . -type f -name DistanceFactory.pm || true
+  - cpanm --notest --force -l . GD
   - cpanm --notest --force -l . CDRAUG/BioPerl-1.7.4.tar.gz
   - cpanm --installdeps --notest .
   - perl Makefile.PL 
