### 2019-12-09 #1762 @lskatz
@csoneson
1. Version information is consistently reported (should I write the tag into the paper?)
2. Tagged release done
3. https://github.com/lskatz/mashtree/releases/tag/v1.0
4. Made a Zenodo archive
5. Zenodo DOI: DOI: 10.5281/zenodo.3568322
