### 2019-12-09 #1762 @csoneson
@lskatz - thanks for this! I'm handing over to the AEiC on duty for the final steps.

@openjournals/joss-eics - I believe we are ready here!
