### 2019-12-10 #1762 @kyleniemeyer
LGTM.
